# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TTS Bot API",
    "version": "1.0.0",
    "description": "API servidor a servidor para que el panel envíe TTS privado, consulte trabajos y administre el overlay."
  },
  "servers": [{ "url": "http://localhost:3000", "description": "Servidor local" }],
  "paths": {
    "/api/cola": {
      "get": {
        "summary": "Obtener cola de mensajes",
        "description": "Devuelve estado actual de la cola",
        "responses": {
          "200": {
            "description": "Cola actual",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": { "type": "integer", "example": 2 },
                    "mensajes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "example": "abc123" },
                          "usuario": { "type": "string", "example": "juan" },
                          "mensaje": { "type": "string", "example": "hola" },
                          "rutaAudio": { "type": "string", "nullable": true }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cola": {
      "get": {
        "summary": "Obtener cola (alias)",
        "description": "Acceso legacy a cola",
        "responses": {
          "200": {
            "description": "Cola actual",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": { "type": "integer", "example": 2 },
                    "mensajes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "example": "abc123" },
                          "usuario": { "type": "string", "example": "juan" },
                          "mensaje": { "type": "string", "example": "hola" },
                          "rutaAudio": { "type": "string", "nullable": true }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Limpiar cola",
        "description": "Vacía la cola y elimina entradas",
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "mensaje": { "type": "string", "example": "Cola limpiada" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/capabilities": {
      "get": {
        "summary": "Obtener capacidades del TTS para construir el panel",
        "security": [{ "PanelToken": [] }],
        "responses": {
          "200": {
            "description": "Límites y estado de autenticación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "apiVersion": { "type": "string", "example": "1.0" },
                    "maxName": { "type": "integer" },
                    "maxMessage": { "type": "integer" },
                    "cooldownSeconds": { "type": "integer" },
                    "actions": {
                      "type": "array",
                      "items": { "type": "string", "enum": ["speak", "ask"] }
                    },
                    "voices": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["id", "label"],
                        "properties": {
                          "id": { "type": "string", "example": "auto" },
                          "label": { "type": "string", "example": "Gemini costeña" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/messages": {
      "post": {
        "summary": "Enviar un mensaje o pregunta desde el panel a OBS",
        "description": "El backend del panel ya validó al usuario. El bot aplica sus límites, genera la voz y encola exactamente una vez por Idempotency-Key.",
        "security": [{ "PanelToken": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string", "minLength": 8, "maxLength": 128 },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["actorId", "name", "message"],
                "properties": {
                  "actorId": { "type": "string", "example": "twitch:123456789" },
                  "name": { "type": "string", "example": "Invitado secreto" },
                  "message": { "type": "string", "example": "Este texto se escucha en OBS" },
                  "voice": {
                    "type": "string",
                    "enum": ["auto", "gemini", "diomedes", "naruto", "google"],
                    "default": "auto",
                    "example": "naruto"
                  },
                  "action": {
                    "type": "string",
                    "enum": ["speak", "ask"],
                    "default": "speak",
                    "example": "speak"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": { "description": "Mensaje agregado o reintento idempotente del mismo trabajo" },
          "400": { "description": "Campos inválidos" },
          "401": { "description": "Credencial del panel incorrecta" },
          "409": { "description": "Idempotency-Key reutilizada con otro contenido" },
          "429": { "description": "Cooldown, presupuesto agotado o cola llena" },
          "503": { "description": "La voz seleccionada no está configurada" }
        }
      }
    },
    "/api/v1/chat/messages/{id}": {
      "get": {
        "summary": "Consultar el estado de un envío reciente",
        "security": [{ "PanelToken": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Estado: generando, listo o finalized" },
          "401": { "description": "Credencial del panel incorrecta" },
          "404": { "description": "Trabajo inexistente o fuera del historial" }
        }
      }
    },
    "/api/config": {
      "get": {
        "summary": "Obtener configuración",
        "description": "Devuelve los valores actuales de UI + TTS",
        "responses": {
          "200": {
            "description": "Configuración actual",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "textColor": { "type": "string", "example": "#eb8352" },
                    "messageColor": { "type": "string", "example": "#f0f0f0" },
                    "background": { "type": "string", "example": "rgba(10, 10, 20, 0.92)" },
                    "borderColor": { "type": "string", "example": "#ea8352" },
                    "barGradient": { "type": "string", "example": "linear-gradient(90deg, #eb8352, #ea8352)" },
                    "voiceLang": { "type": "string", "example": "es-ES" },
                    "voiceRate": { "type": "number", "example": 1.05 },
                    "voicePitch": { "type": "number", "example": 1 }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Actualizar configuración UI/TTS",
        "description": "Guarda nuevo estilo de notificación y parámetros de voz",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "textColor": { "type": "string", "example": "#ff00ff" },
                  "messageColor": { "type": "string", "example": "#ffffff" },
                  "background": { "type": "string", "example": "#0f0f1f" },
                  "borderColor": { "type": "string", "example": "#00ffff" },
                  "barGradient": { "type": "string", "example": "linear-gradient(90deg, #ff00ff, #00ffff)" },
                  "voiceLang": { "type": "string", "example": "es-ES" },
                  "voiceRate": { "type": "number", "format": "float", "minimum": 0.5, "maximum": 2.0, "example": 1.2 },
                  "voicePitch": { "type": "number", "format": "float", "minimum": 0.0, "maximum": 2.0, "example": 1.1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuración actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "config": {
                      "type": "object",
                      "properties": {
                        "textColor": { "type": "string" },
                        "messageColor": { "type": "string" },
                        "background": { "type": "string" },
                        "borderColor": { "type": "string" },
                        "barGradient": { "type": "string" },
                        "voiceLang": { "type": "string" },
                        "voiceRate": { "type": "number" },
                        "voicePitch": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "PanelToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Panel-Token",
        "description": "Secreto exclusivo entre el backend del panel y el bot. También se acepta como Bearer token."
      }
    }
  }
}